Refactored Occi::Api::Dsl into modules
* Refactoring
* Minor changes in option handling
* Connect now accepts a block with configuration
